api/requirements.txt
aniso8601==6.0.0
appdirs==1.4.3
aspy.yaml==1.2.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
cfgv==1.6.0
Click==7.0
coverage==4.5.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.18.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
identify==1.4.1
importlib-metadata==0.9
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
more-itertools==7.0.0
nodeenv==1.3.3
pluggy==0.11.0
pre-commit==1.15.2
psycopg2-binary==2.8.2
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.2
pytest==4.5.0
pytest-cov==2.7.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.0
virtualenv==16.4.3
wcwidth==0.1.7
Werkzeug==0.15.2
zipp==0.3.3
